AdaptiveController_2015/6/16
The 7th submission

Define some macros in the header file Macros.h.
Merge the similar functions in each class.

+++ AdaptiveController/ImageProcessing_Interface.cpp
@@ -185,7 +185,7 @@ void ImageProcessing::setNormalization(IplImage *image, int maxValueIdeal, int m
 		}
 	}
 
-	int range = maxValueCurrent - minValueCurrent;
+	double range = maxValueCurrent - minValueCurrent;
 
 	if(0 == range)
 	{
